Move `Checkpoints.Policy` to `network-layer`
