+++ lib/primitive/cardano-wallet-primitive.cabal
@@ -154,7 +154,6 @@ library
     Cardano.Wallet.Primitive.Types.Block.Gen
     Cardano.Wallet.Primitive.Types.BlockSummary
     Cardano.Wallet.Primitive.Types.Certificates
-    Cardano.Wallet.Primitive.Types.Checkpoints.Policy
     Cardano.Wallet.Primitive.Types.Coin
     Cardano.Wallet.Primitive.Types.Coin.Gen
     Cardano.Wallet.Primitive.Types.DecentralizationLevel
