+++ lib/wallet/src/Cardano/Wallet/Pools.hs
@@ -292,7 +292,7 @@ import UnliftIO.STM
 
 import qualified Cardano.Pool.DB as PoolDb
 import qualified Cardano.Pool.DB.Layer as Pool
-import qualified Cardano.Wallet.Primitive.Types.Checkpoints.Policy as CP
+import qualified Cardano.Wallet.Network.Checkpoints.Policy as CP
 import qualified Cardano.Wallet.Read as Read
 import qualified Data.List as L
 import qualified Data.List.NonEmpty as NE
